switch balance for NewBalance
