+++ App.js
@@ -24,7 +24,7 @@ import ForgotPassword from './src/scenes/ForgotPassword'
 import NewPassword from './src/scenes/ForgotPassword/NewPassword'
 import SetPublicKey from './src/scenes/SetPublicKey'
 import HomeScene from './src/scenes/Home'
-import BalanceScene from './src/scenes/Balance/NewBalance'
+import BalanceScene from './src/scenes/Balance'
 import VoteScene from './src/scenes/Vote'
 import ReceiveScene from './src/scenes/Receive'
 import TransactionListScene from './src/scenes/Transactions'
